# static/openapi.yaml
openapi: 3.1.0
info:
  title: IoT and IM Microservices API (Synchronous)
  description: |
    Formal API specification for the Synchronous Microservices Architecture involving:
    - **IM Services**: Telegram (/IMTelegram) and Line (/IMLine)
    - **IoT Broker**: Central logic and message routing (/IoTBroker)
    - **IoT Device Controller**: Proxy for physical devices (/IoTDevice)
    
    This API facilitates synchronous HTTP communication between chat platforms and IoT devices.
  version: 2.1.0
  contact:
    name: Rocer52
    email: batte9306@gmail.com
servers:
  - url: http://localhost:5000
    description: IM Telegram Service
  - url: http://localhost:5001
    description: IM Line Service
  - url: http://localhost:5002
    description: IoT Broker Service
  - url: http://localhost:5003
    description: IoT Device Controller Service
  - url: http://localhost:5010
    description: ESP32 Virtual Device
  - url: http://localhost:5011
    description: Raspberry Pi Virtual Device
...
paths:
...
  /IMTelegram/send_message:
    post:
      tags:
        - Telegram
      summary: Send Telegram Message
      description: Send a text message to a specific Telegram chat.
      operationId: sendTelegramMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chat_id:
                  type: string
                  example: "123456789"
                text:
                  type: string
                  example: "Hello form System"
              required:
                - chat_id
                - text
      responses:
        '200':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'

  # ==============================
  # Line Microservice Paths
  # ==============================
...
  /IMLine/send_message:
    post:
      tags:
        - Line
      summary: Send Line Message
      description: Send a text message to a specific Line user.
      operationId: sendLineMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chat_id:
                  type: string
                  description: Line User ID
                  example: "U1234567890abcdef1234567890abcdef"
                text:
                  type: string
                  example: "Hello from System"
              required:
                - chat_id
                - text
      responses:
        '200':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'

  # ==============================
  # IoT Broker Paths
  # ==============================
  /IoTBroker/parse_message:
    post:
      tags:
        - IoTBroker
      summary: Parse and Route Message
      description: Analyze incoming text commands and route them to the appropriate device controller.
      operationId: parseIoTMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message_text:
                  type: string
                  example: "turn on esp32_light_001"
                chat_id:
                  type: string
                  example: "123456789"
                platform:
                  type: string
                  enum: [telegram, line]
                  example: "telegram"
                user_id:
                  type: string
                  example: "123456789"
                username:
                  type: string
                  example: "user123"
              required:
                - message_text
                - chat_id
                - platform
      responses:
        '200':
          description: Action executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrokerResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /IoTBroker/bind:
    post:
      tags:
        - IoTBroker
      summary: Bind Device
      description: Bind a user chat ID to a specific device ID.
      operationId: bindDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                device_id:
                  type: string
                  example: "esp32_light_001"
                chat_id:
                  type: string
                  example: "123456789"
                platform:
                  type: string
                  enum: [telegram, line]
                  example: "telegram"
              required:
                - device_id
                - chat_id
                - platform
      responses:
        '200':
          description: Binding successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'
        '400':
          description: Invalid parameters or device ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==============================
  # IoT Device Controller Paths
  # ==============================
...
components:
  schemas:
    BasicResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
        success:
          type: boolean
          example: true
          description: "Alternative key for ok used in some endpoints"
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Invalid request or internal error"
      required:
        - message
    BrokerResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        action:
          type: string
          example: "Enable"
        device_id:
          type: string
          example: "esp32_light_001"
        message:
          type: string
          example: "Command executed"
...